Clean up dgraph functions
